image/cv/requirements/segmentation-lab.txt
segment-geospatial>=0.12.0
rasterio>=1.3.0
geopandas>=1.0.0
shapely>=2.0.0
pyproj>=3.6.0
matplotlib>=3.8.0
pillow>=10.3.0
numpy>=1.26.0
